Feature auth

+++ src/main/java/com/pegazuls/aerodesign/PegStock/service/MaterialService.java
@@ -6,7 +6,6 @@ import java.util.List;
 import com.pegazuls.aerodesign.PegStock.model.dto.borrowing.DTOBorrowingDetails;
 import com.pegazuls.aerodesign.PegStock.model.dto.material.DTOMaterialExpirationDate;
 import com.pegazuls.aerodesign.PegStock.model.dto.material.DTOMaterialMostConsumer;
-import com.pegazuls.aerodesign.PegStock.model.dto.material.DTOMostAvailableMaterial;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -30,6 +29,7 @@ public class MaterialService {
    @Transactional
    public Material create(Material material) {
       validations.forEach(v -> v.validate(material));
+      material.setRegisterDate(LocalDate.now());
       return materialRepository.save(material);
    }
 
@@ -50,9 +50,13 @@ public class MaterialService {
 
    // Update product
    @Transactional
-   public void update(Material material, Long id) {
+   public Material update(Material material, Long id) {
       validations.forEach(v -> v.validate(material));
       Material materialUpdate = materialRepository.findById(id).orElse(null);
+
+      if (materialUpdate == null) {
+         return null; // Product not found
+      }
 
       materialUpdate.setName(material.getName());
       materialUpdate.setDescription(material.getDescription());
@@ -66,11 +70,17 @@ public class MaterialService {
       materialUpdate.setLastAddDate(material.getLastAddDate());
       materialUpdate.setBrand(material.getBrand());
       materialUpdate.setLastConsumptionDate(material.getLastConsumptionDate());
+
+      return materialUpdate;
    }
 
    // Delete product
-   public void delete(Long id) {
-      materialRepository.deleteById(id);
+   public boolean delete(Long id) {
+      if (existsById(id)) {
+         materialRepository.deleteById(id);
+         return true; // Deletion successful
+      }
+      return false; // Material not found
    }
 
    // Check if product exists
@@ -91,7 +101,7 @@ public class MaterialService {
    }
 
    // Method to verify most available product
-   public DTOMostAvailableMaterial mostAvailable() {
+   public DTOMaterialMostConsumer mostAvailable() {
       List<Material> materials = materialRepository.findAll();
       Material material = materials.get(0);
 
@@ -101,7 +111,7 @@ public class MaterialService {
          }
       }
 
-      return new DTOMostAvailableMaterial(material);
+      return new DTOMaterialMostConsumer(material);
    }
 
    // Method to verify nearest expiration product
@@ -115,7 +125,7 @@ public class MaterialService {
          }
       }
 
-      return material == null ? new DTOMaterialExpirationDate("N/A", null) : new DTOMaterialExpirationDate(material);
+      return new DTOMaterialExpirationDate(material);
    }
 
    // List products by category
